+++ src/psa/threshold_union.cpp
@@ -8,10 +8,10 @@ namespace fulgor {
 template <typename T>
 struct scored {
     T item;
-    uint32_t score;
+    uint64_t score;
 };
 
-typedef scored<uint32_t> scored_id;
+typedef scored<uint64_t> scored_id;
 
 template <typename Iterator>
 void merge(std::vector<Iterator>& iterators, std::vector<uint32_t>& colors,
